move cookie utils to their own file
